Anime-Qna/cho/requirements.txt
openai>=1.0.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
chromadb>=0.4.22
opencv-python>=4.8.0
Pillow>=10.0.0
pytesseract>=0.3.10
faster-whisper>=0.10.0
easyocr>=1.7.1
numpy>=1.24.0
tqdm>=4.65.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
sentence-transformers>=2.2.2
